pyproject.toml
[project]
name = "automedia"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=25.1.0",
    "aiosqlite==0.20.0",
    "anthropic==0.28.0",
    "edge-tts>=7.2.7",
    "fastapi==0.111.0",
    "greenlet==3.0.3",
    "httpx==0.27.0",
    "openai==1.30.0",
    "pydantic==2.7.1",
    "pydantic-settings==2.2.1",
    "python-ulid==2.2.0",
    "sqlalchemy==2.0.30",
    "uvicorn[standard]==0.29.0",
    "zhipuai==2.1.5.20250801",
]
